# ops/surfaces/openapi.json
{
  "components": {
    "schemas": {
...
      "BrainstormItem": {
        "properties": {
          "id": {
            "description": "Identifier of the item, unique within the list. Any string.",
            "title": "Id",
            "type": "string"
          },
          "text": {
            "description": "One idea, or one organized step, as free text.",
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "id",
          "text"
        ],
        "title": "BrainstormItem",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "InboxItem": {
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "priority": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority"
          }
        },
        "required": [
          "description"
        ],
        "title": "InboxItem",
        "type": "object"
      },
...
      "ProjectCreate": {
        "properties": {
          "name": {
            "description": "Project name, spelled exactly as tasks spell it; dots nest subprojects (`home.garden`). Surrounding whitespace is removed.",
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "ProjectCreate",
        "type": "object"
      },
      "ProjectPlan": {
        "properties": {
          "brainstorm": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/BrainstormItem"
            },
            "title": "Brainstorm",
            "type": "array"
          },
          "organized": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/BrainstormItem"
            },
            "title": "Organized",
            "type": "array"
          },
          "principles": {
            "default": "",
            "title": "Principles",
            "type": "string"
          },
          "project_name": {
            "title": "Project Name",
            "type": "string"
          },
          "purpose": {
            "default": "",
            "title": "Purpose",
            "type": "string"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "vision": {
            "default": "",
            "title": "Vision",
            "type": "string"
          }
        },
        "required": [
          "project_name"
        ],
        "title": "ProjectPlan",
        "type": "object"
      },
...
      "Task": {
        "properties": {
          "annotations": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/TaskAnnotation"
            },
            "title": "Annotations",
            "type": "array"
          },
          "depends": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Depends",
            "type": "array"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "due": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due"
          },
          "end": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End"
          },
          "entry": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entry"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "modified": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Modified"
          },
          "priority": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority"
          },
          "project": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project"
          },
          "recur": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recur"
          },
          "scheduled": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduled"
          },
          "start": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "tags": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "until": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Until"
          },
          "urgency": {
            "default": 0.0,
            "title": "Urgency",
            "type": "number"
          },
          "uuid": {
            "title": "Uuid",
            "type": "string"
          },
          "wait": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wait"
          }
        },
        "required": [
          "uuid",
          "id",
          "description",
          "status"
        ],
        "title": "Task",
        "type": "object"
      },
      "TaskAnnotation": {
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "entry": {
            "title": "Entry",
            "type": "string"
          }
        },
        "required": [
          "entry",
          "description"
        ],
        "title": "TaskAnnotation",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "A self-hosted GTD task manager. Authenticate with JWT (Authorization: Bearer <token>) or API key (X-Api-Key: <key>).",
    "title": "Runway",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/gtd/inbox": {
      "get": {
        "description": "List tasks that have not been processed yet: no project and no tags. A tag means the task has been clarified, so giving it one takes it out of the inbox.",
        "operationId": "inbox_gtd_inbox_get",
        "parameters": [
          {
            "description": "Only tasks carrying ALL of these tags, written without `+`. Repeat the parameter for more than one. The inbox is untagged by definition, so any tag yields an empty inbox.",
            "in": "query",
            "name": "tag",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "maxItems": 10,
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only tasks carrying ALL of these tags, written without `+`. Repeat the parameter for more than one. The inbox is untagged by definition, so any tag yields an empty inbox.",
              "title": "Tag"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Task"
                  },
                  "title": "Response Inbox Gtd Inbox Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "GTD inbox",
        "tags": [
          "gtd"
        ]
      }
    },
    "/gtd/next": {
      "get": {
        "description": "List tasks tagged +next \u2014 concrete actions you can do right now.",
        "operationId": "next_actions_gtd_next_get",
        "parameters": [
          {
            "description": "Only tasks carrying ALL of these tags, written without `+`. Repeat the parameter for more than one. The inbox is untagged by definition, so any tag yields an empty inbox.",
            "in": "query",
            "name": "tag",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "maxItems": 10,
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only tasks carrying ALL of these tags, written without `+`. Repeat the parameter for more than one. The inbox is untagged by definition, so any tag yields an empty inbox.",
              "title": "Tag"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Task"
                  },
                  "title": "Response Next Actions Gtd Next Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Next actions",
        "tags": [
          "gtd"
        ]
      }
    },
    "/gtd/projects": {
      "get": {
        "description": "Return all project names \u2014 both those inferred from tasks and those created explicitly.",
        "operationId": "projects_gtd_projects_get",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "type": "string"
                  },
                  "title": "Response Projects Gtd Projects Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List projects",
        "tags": [
          "gtd"
        ]
      }
    },
    "/gtd/projects/{name}": {
      "get": {
        "description": "List all pending tasks belonging to a specific project.",
        "operationId": "project_tasks_gtd_projects__name__get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          },
          {
            "description": "Only tasks carrying ALL of these tags, written without `+`. Repeat the parameter for more than one. The inbox is untagged by definition, so any tag yields an empty inbox.",
            "in": "query",
            "name": "tag",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "maxItems": 10,
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only tasks carrying ALL of these tags, written without `+`. Repeat the parameter for more than one. The inbox is untagged by definition, so any tag yields an empty inbox.",
              "title": "Tag"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Task"
                  },
                  "title": "Response Project Tasks Gtd Projects  Name  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Project tasks",
        "tags": [
          "gtd"
        ]
      }
    },
    "/gtd/someday": {
      "get": {
        "description": "List tasks tagged +someday \u2014 ideas and intentions not yet committed to.",
        "operationId": "someday_gtd_someday_get",
        "parameters": [
          {
            "description": "Only tasks carrying ALL of these tags, written without `+`. Repeat the parameter for more than one. The inbox is untagged by definition, so any tag yields an empty inbox.",
            "in": "query",
            "name": "tag",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "maxItems": 10,
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only tasks carrying ALL of these tags, written without `+`. Repeat the parameter for more than one. The inbox is untagged by definition, so any tag yields an empty inbox.",
              "title": "Tag"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Task"
                  },
                  "title": "Response Someday Gtd Someday Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Someday / maybe",
        "tags": [
          "gtd"
        ]
      }
    },
    "/gtd/tickler": {
      "get": {
        "description": "Tasks hidden by a future wait date, soonest first. They return to their list (untagged ones to the inbox) when the date passes.",
        "operationId": "tickler_gtd_tickler_get",
        "parameters": [
          {
            "description": "Only tasks carrying ALL of these tags, written without `+`. Repeat the parameter for more than one. The inbox is untagged by definition, so any tag yields an empty inbox.",
            "in": "query",
            "name": "tag",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "maxItems": 10,
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only tasks carrying ALL of these tags, written without `+`. Repeat the parameter for more than one. The inbox is untagged by definition, so any tag yields an empty inbox.",
              "title": "Tag"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Task"
                  },
                  "title": "Response Tickler Gtd Tickler Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Tickler",
        "tags": [
          "gtd"
        ]
      }
    },
    "/gtd/waiting": {
      "get": {
        "description": "List tasks tagged +waiting \u2014 things delegated or blocked on someone/something else.",
        "operationId": "waiting_gtd_waiting_get",
        "parameters": [
          {
            "description": "Only tasks carrying ALL of these tags, written without `+`. Repeat the parameter for more than one. The inbox is untagged by definition, so any tag yields an empty inbox.",
            "in": "query",
            "name": "tag",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "maxItems": 10,
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only tasks carrying ALL of these tags, written without `+`. Repeat the parameter for more than one. The inbox is untagged by definition, so any tag yields an empty inbox.",
              "title": "Tag"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Task"
                  },
                  "title": "Response Waiting Gtd Waiting Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Waiting for",
        "tags": [
          "gtd"
        ]
      }
    },
...
    "/inbox": {
      "post": {
        "description": "Add a task to the inbox of the authenticated user. Accepts the same credentials as every other route \u2014 'X-Api-Key: <key>' or 'Authorization: Bearer <jwt>' \u2014 and, for compatibility with existing agents, an API key in the Bearer slot. Intended for agents and automations.",
        "operationId": "webhook_inbox_inbox_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InboxItem"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Add to inbox via API key",
        "tags": [
          "inbox"
        ]
      }
    },
    "/projects": {
      "post": {
        "description": "Create a project explicitly (it also exists implicitly as soon as a task names it). Creating the same name twice changes nothing and returns the stored plan.",
        "operationId": "create_project_projects_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectPlan"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create a project",
        "tags": [
          "projects"
        ]
      }
    },
...
  }
}